refactor(web/applications): move case service to higher level
