+++ apps/web/.env.test
@@ -64,7 +64,7 @@ LOG_LEVEL_FILE=debug
 
 # The logging level for the stdout.
 # Options: trace, debug, info, warn, error, fatal, silent
-LOG_LEVEL_STDOUT=debug
+LOG_LEVEL_STDOUT=fatal
 
 ## SERVICE COMMON  ####################################################################
