+++ apps/web/src/server/applications/lib/controllers/errors.controller.js
@@ -1,0 +1,16 @@
+/** @typedef {import('../../applications.types.js').FormCaseLayout} FormCaseLayout */
+
+/**
+ * View the sector choice step of the application creation
+ *
+ * @param {Record<string, any>} properties
+ * @param {FormCaseLayout} layout
+ * @param {*} response
+ * @returns {*}
+ */
+export const handleErrors = (properties, layout, response) => {
+	return response.render('applications/components/case-form/case-form-layout', {
+		...properties,
+		layout
+	});
+};
