+++ apps/web/src/server/applications/lib/services/session.service.js
@@ -1,7 +1,43 @@
 /**
  * @typedef {import('express-session').Session & { caseSectorName?: string }} SessionWithCaseSectorName
  * @typedef {import('express-session').Session & { caseHasNeverBeenResumed?: boolean }} SessionWithCaseHasNeverBeenResumed
+ * @typedef {import('express-session').Session & { infoTypes?: string[] }} SessionWithApplicationsCreateApplicantInfoTypes
  */
+
+// Applicant session management
+
+/**
+ * Save in the session the list of information types to be provided in the create-new-applicant form.
+ *
+ * @param {SessionWithApplicationsCreateApplicantInfoTypes} session
+ * @param {string[]} infoTypes
+ * @returns {void}
+ */
+export const setSessionApplicantInfoTypes = (session, infoTypes) => {
+	session.infoTypes = ['applicant-information-types', ...infoTypes];
+};
+
+/**
+ * Retrieve the applicant information types from the session.
+ *
+ * @param {SessionWithApplicationsCreateApplicantInfoTypes} session
+ * @returns {string[]}
+ */
+export const getSessionApplicantInfoTypes = (session) => {
+	return session.infoTypes ?? [];
+};
+
+/**
+ * Clear the list of applicant information types
+ *
+ * @param {SessionWithApplicationsCreateApplicantInfoTypes} session
+ * @returns {void}
+ */
+export const destroySessionApplicantInfoTypes = (session) => {
+	delete session.infoTypes;
+};
+
+// Application session management
 
 /**
  * Save in the session the list of information types to be provided in the create-new-applicant form.
