+++ apps/web/src/server/applications/lib/services/utilities.js
@@ -1,9 +1,9 @@
-import { get } from '../../lib/request.js';
+import { get } from '../../../lib/request.js';
 
-/** @typedef {import('../applications.types').Sector} Sector */
-/** @typedef {import('../applications.types').Application} Application */
-/** @typedef {import('../applications.types').Region} Region */
-/** @typedef {import('../applications.types.js').ZoomLevel} ZoomLevel */
+/** @typedef {import('../../applications.types').Sector} Sector */
+/** @typedef {import('../../applications.types').Application} Application */
+/** @typedef {import('../../applications.types').Region} Region */
+/** @typedef {import('../../applications.types').ZoomLevel} ZoomLevel */
 
 /**
  * Get the list of sector for an application
